Fix/remove python 3.7 (#536)
* add python 3.11 to CI workflow
* drop support for python 3.7
* update the libstd for the CI runner to fix bugs related to newer scipy versions
* fix dtype error appearing with quantities 0.14.0

+++ elephant/test/test_conversion.py
@@ -510,8 +510,8 @@ class BinnedSpikeTrainTestCase(unittest.TestCase):
     def test_to_array(self):
         x = cv.BinnedSpikeTrain(self.spiketrain_a, bin_size=1 * pq.s,
                                 n_bins=10, t_stop=10. * pq.s)
-        arr_float = x.to_array(dtype=np.float32)
-        assert_array_equal(arr_float, x.to_array().astype(np.float32))
+        arr_float = x.to_array(dtype=float)
+        assert_array_equal(arr_float, x.to_array().astype(float))
 
     # Test if error is raised when providing insufficient number of
     # parameters
